// frontend/react-frontend-a_5/src/models/FullUser.ts
export {};

export interface FullUser{
    username: string;
    user_first_name: string;
    user_last_name: string;
    user_date_of_birth: string;
    user_bio: string;
    user_location: string;
    swimmers_count:number;
    teams_count:number;
    coaches_count:number;
    fans_count:number;
};